Render distribution heatmaps as a magma density raster and restructure the metric timeline
The heatmap now builds a column-normalized density grid (one row per bin
step, capped at plot-pixel resolution), maps it through a magma lookup
table, and paints it with a single drawImage — replacing one fillRect and
one fillStyle string per bin, which took tens of milliseconds per redraw
at high run counts and saturated to black. Rendering is fixed: raster,
magma, per-column normalization, no smoothing.

The component splits into a private subtree — shared frame types and
distribution math, plot data and options builders, the heatmap plugin, a
self-positioning frame popover, and a useMetricPlot hook owning the uPlot
lifecycle — declared as the ui.views.editor.metric-timeline arch-docs
layer. useElementSize gains a border-box option so popover placement
measures the element's real size. The dev-only heatmap lab is removed;
its pipeline tests carry over to the heatmap module.
