+++ .changeset/heatmap-raster-magma.md
@@ -1,0 +1,5 @@
+---
+"@hashintel/petrinaut": patch
+---
+
+Distribution heatmaps in experiment metric charts render as a magma-colored density image — one draw call instead of a rectangle per bin — making them fast at high run counts and readable across the full spread of runs.
